--- client/src/components/organisms/form/UserEditForm.tsx
@@ -1,260 +1,0 @@
-import React from 'react';
-import * as yup from 'yup';
-import styled from 'styled-components';
-import { useFormik } from 'formik';
-import { Padding, ALIGNITEMS, JUSTIFYCONTENT } from '../../../utils/styled/styledSpace';
-import { media } from '../../../utils/styled/styledRdesign';
-import { FONTSIZE, SIZE, CLEAR } from '../../../utils/constant/number';
-import { BASICCOLORS } from '../../../utils/constant/color';
-import FlexLayout from '../../atoms/FlexLayout';
-import { baseUser } from '../../../utils/constant/text/body/user/value';
-import { selectProfileItems } from '../../pages/UserEdit';
-import FormTitle from '../../atoms/form/FormTitle';
-import FormSubmit from '../../atoms/form/FormSubmit';
-import { emailValidation, nameValidation } from '../../../validations';
-
-type SelectItems = typeof selectProfileItems | typeof selectGolfItems;
-type EditItems = typeof baseItems | typeof snsItems | typeof golfItems | typeof otherItems;
-type NoteItems = typeof snsNoteItems | typeof profileNoteItems;
-
-interface Props {
-  currentUser: UserType;
-  onSubmit: (values: ProfileEditDataType) => void;
-}
-
-const StyledForm = styled.form``;
-
-const StyledField = styled.input`
-  width: ${SIZE.SMALL}vw;
-  font-size: ${FONTSIZE.LARGE}px;
-  padding: ${CLEAR.TINY}vw 0px;
-  border-radius: 6px;
-  border-width: 1px;
-  border: 1px solid #ccc;
-  ${media.tablet`
-      width: ${SIZE.MEDIUMLARGE}vw;
-      `}
-`;
-const StyledSelect = styled.select`
-  width: ${SIZE.XXXSMALL}vw;
-  font-size: ${FONTSIZE.MEDIUM}px;
-  padding: ${CLEAR.TINY}vw 0px;
-  border-radius: 6px;
-  border-width: 1px;
-  border: 1px solid #ccc;
-  ${media.tablet`
-      width: ${SIZE.MEDIUMLARGE}vw;
-      `}
-`;
-
-const Styleddiv = styled.div`
-  margin: 0 auto;
-  color: ${BASICCOLORS.SECONDARYDARK};
-`;
-
-const Inline = styled.div`
-  font-size: ${FONTSIZE.SMALL}px;
-`;
-const StyledLabel = styled.label`
-  font-size: ${FONTSIZE.BASE}px;
-  color: ${BASICCOLORS.BASICDARK};
-`;
-
-const baseItems = {
-  name: '名前',
-  email: 'メールアドレス',
-};
-const snsItems = {
-  twitter: 'twitterアカウント',
-  instagram: 'instagramアカウント',
-  youtube: 'youtubeアカウント',
-  facebook: 'facebookアカウント',
-};
-const golfItems = {
-  homeCourse: 'ホームコース',
-};
-const otherItems = {
-  residence: '現住所',
-  birthPlace: '出生地',
-  job: '職業',
-  school: '出身校',
-  hobby: '趣味',
-};
-let iniScore = 54;
-const bScore = [...Array(82)].map((_, i) => {
-  return iniScore + i;
-});
-let iniDistance = 150;
-const aDistance = [...Array(21)].map((_, i) => {
-  return iniDistance + i * 10;
-});
-const selectGolfItems = {
-  bestScore: { head: 'ベストスコア', body: bScore },
-  averageDistance: { head: '平均飛距離', body: aDistance },
-};
-
-const snsNoteItems = {
-  twitter: 'twitter.com/後のアカウント名。',
-  instagram: 'www.instagram.com/後のアカウント名。',
-  youtube: 'www.youtube.com/channel/後のアカウント名。',
-  facebook: 'www.facebook.com/後のアカウント名。',
-};
-const profileNoteItems = {
-  show: 'ログインユーザーのみに情報が公開されます。',
-};
-
-const editTitles = ['基本情報', 'ゴルフ', 'SNS', 'その他'];
-
-const buttonValue = 'プロフィールを編集する。';
-
-const profileValidation = () =>
-  yup.object().shape({
-    email: emailValidation(),
-    name: nameValidation(),
-  });
-
-const UserEditForm: React.FC<Props> = ({ currentUser, onSubmit }) => {
-  const showValue = currentUser.show
-    ? selectProfileItems.show.body[0]
-    : selectProfileItems.show.body[1];
-  const editFormDatas = {
-    initialValuesData: {
-      ...currentUser,
-      show: showValue,
-      confirmedPassword: currentUser.password,
-    },
-    placeHolder: {
-      ...baseUser,
-      password: '英数字８字以上のパスワード',
-      confirmedPassword: '確認用パスワード',
-    },
-  };
-  const formik = useFormik({
-    initialValues: { ...editFormDatas.initialValuesData },
-    validationSchema: profileValidation,
-    onSubmit: onSubmit,
-  });
-
-  const selectItems = (obj: SelectItems, note?: NoteItems) => {
-    const element = Object.entries(obj).map(([key, value], num: number) => {
-      const { head, body } = value;
-      return (
-        <React.Fragment key={num}>
-          <Padding top={CLEAR.TINY} bottom={CLEAR.TINY}>
-            <FlexLayout
-              justifyContent={JUSTIFYCONTENT.START}
-              width={SIZE.XXXSMALL}
-              alignItems={ALIGNITEMS.START}
-              left={
-                <Padding left={CLEAR.TINY}>
-                  <StyledLabel htmlFor={key}>{head}</StyledLabel>
-                </Padding>
-              }
-              right={
-                <Padding left={CLEAR.MEDIUM}>
-                  <StyledSelect name={key} onChange={formik.handleChange}>
-                    <option value={formik.values[key as keyof UserType]}>
-                      {formik.values[key as keyof UserType]}
-                    </option>
-                    {body.map((data: any, num: number) => {
-                      return formik.values[key as keyof UserType] !== data ? (
-                        <option key={num} value={data}>
-                          {data}
-                        </option>
-                      ) : null;
-                    })}
-                  </StyledSelect>
-                  {note === undefined ? <></> : <Inline>{note[key as keyof NoteItems]}</Inline>}
-                </Padding>
-              }
-            />
-          </Padding>
-          {formik.touched[key as keyof UserType] && formik.errors[key as keyof UserType] ? (
-            <Styleddiv>{formik.errors[key as keyof UserType]}</Styleddiv>
-          ) : null}
-        </React.Fragment>
-      );
-    });
-    return element;
-  };
-
-  const InputItems = (obj: EditItems, note?: NoteItems) => {
-    const element = Object.entries(editFormDatas.placeHolder).map(
-      ([key, value]: string[], num: number) => {
-        return obj[key as keyof EditItems] ? (
-          <React.Fragment key={num}>
-            <Padding top={CLEAR.TINY} bottom={CLEAR.TINY}>
-              <FlexLayout
-                justifyContent={JUSTIFYCONTENT.START}
-                width={SIZE.XXXSMALL}
-                alignItems={ALIGNITEMS.START}
-                left={
-                  <Padding left={CLEAR.TINY}>
-                    <StyledLabel htmlFor={key}>{obj[key as keyof EditItems]}</StyledLabel>
-                  </Padding>
-                }
-                right={
-                  <Padding left={CLEAR.MEDIUM}>
-                    <StyledField
-                      type={key}
-                      name={key}
-                      placeholder={value}
-                      onChange={formik.handleChange}
-                      onBlur={formik.handleBlur}
-                      value={formik.values[key as keyof UserType]}
-                    />
-                    {note === undefined ? <></> : <Inline>{note[key as keyof NoteItems]}</Inline>}
-                  </Padding>
-                }
-              />
-            </Padding>
-            {formik.touched[key as keyof UserType] && formik.errors[key as keyof UserType] ? (
-              <Styleddiv>{formik.errors[key as keyof UserType]}</Styleddiv>
-            ) : null}
-          </React.Fragment>
-        ) : (
-          <React.Fragment key={num}></React.Fragment>
-        );
-      }
-    );
-    return element;
-  };
-
-  const editElements = [
-    <>
-      {InputItems(baseItems)}
-      {selectItems(selectProfileItems, profileNoteItems)}
-    </>,
-    <>
-      {InputItems(golfItems)}
-      {selectItems(selectGolfItems)}
-    </>,
-    <>{InputItems(snsItems, snsNoteItems)}</>,
-    <>{InputItems(otherItems)}</>,
-  ];
-
-  const editTitleElemensts = editElements.map((element, i) => {
-    return i === 0 ? (
-      <Padding key={i} top={CLEAR.XSMALL} bottom={CLEAR.SMALL}>
-        <FormTitle>{editTitles[i]}</FormTitle>
-        {element}
-      </Padding>
-    ) : (
-      <Padding key={i} bottom={CLEAR.SMALL}>
-        <FormTitle>{editTitles[i]}</FormTitle>
-        {element}
-      </Padding>
-    );
-  });
-
-  return (
-    <Padding right={CLEAR.MEDIUM} left={CLEAR.MEDIUM}>
-      <StyledForm onSubmit={formik.handleSubmit}>
-        {editTitleElemensts}
-        <FormSubmit>{buttonValue}</FormSubmit>
-      </StyledForm>
-    </Padding>
-  );
-};
-
-export default UserEditForm;
